+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2019/07/15
+- [UPDATE][COOKBOOK] regulus to 2.2.0
+
 # 2019/07/13
 - [UPDATE][COOKBOOK] regulus to 2.1.9